Trabajo 1 rehecho

+++ Variables/cuadratica/equivalencia_dias.vb
@@ -3,9 +3,9 @@ Imports System
 Module equivalencia_dias
     Sub Main()
         Dim numero As Integer
-        Const ds As Int16 = 7
-        Const hd As Int16 = 24
-        Const mh As Int16 = 60
+        Const ds As Byte = 7
+        Const hd As Byte = 24
+        Const mh As Byte = 60
 
         Console.WriteLine("Ingresar un numero")
         numero = Console.ReadLine()
